pyproject.toml
[project]
name = "priompt"
version = "0.1.0"
description = "Priompt (priority + prompt) is prompting library. It uses priorities to decide what to include in the context window."
authors = [{ name = "Cyrus Nouroozi", email = "cyrus@zenbase.ai" }]
dependencies = [
    "tiktoken>=0.8.0",
    "pyyaml>=6.0.2",
    "ujson>=5.10.0",
    "beartype>=0.19.0",
    "typing-extensions>=4.12.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipdb>=0.13.13",
    "ipython>=8.29.0",
    "syrupy>=4.8.0",
    "hypothesis>=6.119.4",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/priompt"]

[tool.ruff]
line-length = 100

[tool.project.optional-dependencies]
dev = ["tox", "tox-gh-actions"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
